+++ CHANGELOG.md
@@ -8,3 +8,4 @@
     * `clearIndices()`
 * `leinonen\Yii2Agolia\AlgoliaManager` and `leinonen\Yii2Algolia\ActiveRecord\Searchable` have now feature to do backend searches with `search()` method.
 * Some of integration tests are now run against a real database
+* All of the base methods are now covered by integration tests.